Extracts Navbar related components and uses CSS Modules
