+++ src/components/common/Navbar/Navbar.tsx
@@ -1,11 +1,8 @@
-import { List as UIList } from '@faststore/ui'
-import { Suspense, useRef, useState } from 'react'
-import type { AnchorHTMLAttributes } from 'react'
 import type { SearchInputRef } from '@faststore/ui'
+import { Suspense, useRef, useState } from 'react'
 
 import CartToggle from 'src/components/cart/CartToggle'
 import SearchInput from 'src/components/common/SearchInput'
-import RegionalizationButton from 'src/components/regionalization/RegionalizationButton'
 import {
   ButtonIcon,
   ButtonSignIn,
@@ -14,97 +11,12 @@ import {
 import Icon from 'src/components/ui/Icon'
 import Link from 'src/components/ui/Link'
 import Logo from 'src/components/ui/Logo'
-import SlideOver from 'src/components/ui/SlideOver'
 import { mark } from 'src/sdk/tests/mark'
 import { useUI } from 'src/sdk/ui/Provider'
-import { useFadeEffect } from 'src/sdk/ui/useFadeEffect'
 
-interface NavLinksProps {
-  onClickLink?: AnchorHTMLAttributes<HTMLAnchorElement>['onClick']
-}
-
-const collections = [
-  {
-    name: 'Office',
-    href: '/office',
-  },
-  {
-    name: 'Home Appliances',
-    href: '/kitchen---home-appliances',
-  },
-  {
-    name: 'Computer and Software',
-    href: '/computer---software',
-  },
-  {
-    name: 'Technology',
-    href: '/technology',
-  },
-]
-
-function NavLinks({ onClickLink }: NavLinksProps) {
-  return (
-    <nav className="navlinks__list">
-      <UIList>
-        <li>
-          <RegionalizationButton classes="hidden-mobile" />
-        </li>
-        {collections.map(({ href, name }) => (
-          <li key={name}>
-            <Link variant="display" href={href} onClick={onClickLink}>
-              {name}
-            </Link>
-          </li>
-        ))}
-      </UIList>
-    </nav>
-  )
-}
-
-function NavbarSlider() {
-  const { closeNavbar } = useUI()
-  const { fade, fadeOut } = useFadeEffect()
-
-  return (
-    <SlideOver
-      isOpen
-      fade={fade}
-      onDismiss={fadeOut}
-      size="full"
-      direction="leftSide"
-      className="navbar__modal-content"
-      onTransitionEnd={() => fade === 'out' && closeNavbar()}
-    >
-      <div className="navbar__modal-body">
-        <header className="navbar__modal-header">
-          <Link
-            href="/"
-            onClick={fadeOut}
-            aria-label="Go to FastStore home"
-            title="Go to FastStore home"
-            className="navbar__logo"
-          >
-            <Logo />
-          </Link>
-
-          <ButtonIcon
-            aria-label="Close Menu"
-            icon={<Icon name="X" width={32} height={32} />}
-            onClick={fadeOut}
-          />
-        </header>
-        <div className="navlinks">
-          <NavLinks onClickLink={fadeOut} />
-          <div className="navlinks__signin">
-            <Suspense fallback={<ButtonSignInFallback />}>
-              <ButtonSignIn />
-            </Suspense>
-          </div>
-        </div>
-      </div>
-    </SlideOver>
-  )
-}
+import styles from './navbar.module.scss'
+import NavbarSlider from './NavbarSlider'
+import NavLinks from './NavLinks'
 
 function Navbar() {
   const { openNavbar, navbar: displayNavbar } = useUI()
@@ -117,13 +29,13 @@ function Navbar() {
   }
 
   return (
-    <header className="navbar layout__content-full">
-      <div className="navbar__header layout__content">
-        <section className="navbar__row">
+    <header className={`${styles.fsNavbar} layout__content-full`}>
+      <div className="layout__content" data-fs-navbar-header>
+        <section data-fs-navbar-row>
           {!searchExpanded && (
             <>
               <ButtonIcon
-                data-fs-button-menu
+                data-fs-navbar-button-menu
                 aria-label="Open Menu"
                 icon={<Icon name="List" width={32} height={32} />}
                 onClick={openNavbar}
@@ -132,7 +44,7 @@ function Navbar() {
                 href="/"
                 aria-label="Go to Faststore home"
                 title="Go to Faststore home"
-                className="navbar__logo"
+                data-fs-navbar-logo
               >
                 <Logo />
               </Link>
@@ -140,7 +52,7 @@ function Navbar() {
           )}
           <SearchInput />
           <div
-            className="navbar__buttons"
+            data-fs-navbar-buttons
             data-store-search-expanded={searchExpanded}
           >
             {searchExpanded && (
@@ -164,7 +76,7 @@ function Navbar() {
             <CartToggle />
           </div>
         </section>
-        <NavLinks />
+        <NavLinks classes="hidden-mobile" />
       </div>
 
       {displayNavbar && <NavbarSlider />}
