+++ src/components/common/Navbar/index.tsx
@@ -1,1 +1,3 @@
 export { default } from './Navbar'
+export { default as NavbarSlider } from './NavbarSlider'
+export { default as NavLinks } from './NavLinks'
