+++ src/components/common/Navbar/navbar-slider.module.scss
@@ -1,0 +1,38 @@
+@import "src/styles/scaffold";
+
+.fs-navbar-slider {
+  padding: var(--fs-spacing-3);
+
+  [data-fs-navbar-slider-header] {
+    position: relative;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding-bottom: var(--fs-spacing-2);
+
+    [data-fs-button-icon] {
+      margin-right: calc(-1 * var(--fs-spacing-1));
+    }
+  }
+
+  [data-fs-navbar-slider-logo] {
+    height: 2.375rem;
+    padding: 0;
+    margin-right: var(--fs-spacing-5);
+
+    img {
+      width: 100%;
+      height: 100%;
+    }
+  }
+
+  [data-fs-navbar-slider-footer] {
+    padding-top: var(--fs-spacing-2);
+    margin-top: var(--fs-spacing-2);
+
+    [data-fs-button-signin-link] {
+      width: fit-content;
+      padding-left: 0;
+    }
+  }
+}
