+++ src/components/common/Navbar/navlinks.module.scss
@@ -1,0 +1,31 @@
+@import "src/styles/scaffold";
+
+.fs-navlinks {
+  li {
+    display: flex;
+    align-items: center;
+  }
+
+  [data-fs-link] {
+    display: block;
+    width: 100%;
+  }
+
+  @include media(">=notebook") {
+    ul {
+      display: flex;
+      column-gap: var(--fs-spacing-3);
+      padding-top: var(--fs-spacing-1);
+    }
+
+    [data-fs-link] {
+      width: auto;
+      padding: var(--fs-spacing-0) 0;
+    }
+  }
+
+  @include media("<notebook") {
+    border-top: var(--fs-border-width) solid var(--fs-color-neutral-2);
+    border-bottom: var(--fs-border-width) solid var(--fs-color-neutral-2);
+  }
+}
