+++ src/styles/global/components.scss
@@ -6,7 +6,6 @@
 // Common
 @import "src/components/common/Footer/footer.scss";
 @import "src/components/common/SearchInput/search-input.scss";
-@import "src/components/common/Navbar/navbar.scss";
 @import "src/components/regionalization/RegionalizationBar/regionalization-bar.scss";
 @import "src/components/regionalization/RegionalizationButton/regionalization-button.scss";
 @import "src/components/regionalization/RegionalizationModal/regionalization-modal.scss";
