+++ src/styles/global/storybook-components.scss
@@ -6,7 +6,6 @@
 // Common
 @import "src/components/common/Footer/footer.scss";
 @import "src/components/common/SearchInput/search-input.scss";
-@import "src/components/common/Navbar/navbar.scss";
 
 // Search
 @import "src/components/search/Filter/filter.scss";
